ekf: merge backend classes EstimatorInterface + Ekf
