+++ src/modules/ekf2/EKF/optical_flow_control.cpp
@@ -38,6 +38,41 @@
 
 #include "ekf.h"
 
+void Ekf::setOpticalFlowData(const flowSample &flow)
+{
+	if (!_initialised) {
+		return;
+	}
+
+	// Allocate the required buffer size if not previously done
+	if (_flow_buffer == nullptr) {
+		_flow_buffer = new RingBuffer<flowSample>(_imu_buffer_length);
+
+		if (_flow_buffer == nullptr || !_flow_buffer->valid()) {
+			delete _flow_buffer;
+			_flow_buffer = nullptr;
+			printBufferAllocationFailed("flow");
+			return;
+		}
+	}
+
+	const int64_t time_us = flow.time_us
+				- static_cast<int64_t>(_params.flow_delay_ms * 1000)
+				- static_cast<int64_t>(_dt_ekf_avg * 5e5f); // seconds to microseconds divided by 2
+
+	// limit data rate to prevent data being lost
+	if (time_us >= static_cast<int64_t>(_flow_buffer->get_newest().time_us + _min_obs_interval_us)) {
+
+		flowSample optflow_sample_new{flow};
+		optflow_sample_new.time_us = time_us;
+
+		_flow_buffer->push(optflow_sample_new);
+
+	} else {
+		ECL_WARN("optical flow data too fast %" PRIi64 " < %" PRIu64 " + %d", time_us, _flow_buffer->get_newest().time_us, _min_obs_interval_us);
+	}
+}
+
 void Ekf::controlOpticalFlowFusion(const imuSample &imu_delayed)
 {
 	if (_flow_buffer) {
