+++ src/modules/ekf2/EKF/range_height_control.cpp
@@ -38,6 +38,42 @@
 
 #include "ekf.h"
 
+void Ekf::setRangeData(const rangeSample &range_sample)
+{
+	if (!_initialised) {
+		return;
+	}
+
+	// Allocate the required buffer size if not previously done
+	if (_range_buffer == nullptr) {
+		_range_buffer = new RingBuffer<rangeSample>(_obs_buffer_length);
+
+		if (_range_buffer == nullptr || !_range_buffer->valid()) {
+			delete _range_buffer;
+			_range_buffer = nullptr;
+			printBufferAllocationFailed("range");
+			return;
+		}
+	}
+
+	const int64_t time_us = range_sample.time_us
+				- static_cast<int64_t>(_params.range_delay_ms * 1000)
+				- static_cast<int64_t>(_dt_ekf_avg * 5e5f); // seconds to microseconds divided by 2
+
+	// limit data rate to prevent data being lost
+	if (time_us >= static_cast<int64_t>(_range_buffer->get_newest().time_us + _min_obs_interval_us)) {
+
+		rangeSample range_sample_new{range_sample};
+		range_sample_new.time_us = time_us;
+
+		_range_buffer->push(range_sample_new);
+		_time_last_range_buffer_push = _time_latest_us;
+
+	} else {
+		ECL_WARN("range data too fast %" PRIi64 " < %" PRIu64 " + %d", time_us, _range_buffer->get_newest().time_us, _min_obs_interval_us);
+	}
+}
+
 void Ekf::controlRangeHeightFusion()
 {
 	static constexpr const char *HGT_SRC_NAME = "RNG";
